refactor(shared-log): extract native write-through block store module
Move the NativeBackboneWriteThroughBlockStore class (index.ts lines
456-1613: NativeBackboneBlocks / NativeCommitOwnershipToken types, class
JSDoc, and the class body) verbatim into
src/native-write-through-block-store.ts. The class was already
self-contained: its only dependencies are the native/durable constructor
parameter properties, so no deps object is needed. The sole textual
change to moved code is adding export.

Module-local glue per the extraction manifest: recreate the warn scope
from the same peerbit:shared-log logger namespace (importing logger from
index.ts would create a cycle; identical debug namespace, zero behavior
change), and copy the index-private MaybePromise type and isPromiseLike
helper (index.ts keeps its originals, which remain heavily used there).

index.ts gains one import of the class; construction in open() and the
instanceof poison-clear gate are the only two references. The class is
deliberately not re-exported from index.ts (not public API). The
createNativeDurableBlockStore / createDefaultDurableBlockStore factories
and DurableBlockSublevelStore type stay in index.ts per the manifest.

Ratchet: remove nativeBlockWriteGenerations and nativeDeleteEpoch from
the check-fence-ratchet.mjs BASELINE since the ratchet targets index.ts
only and both fields moved out. Deviation from the manifest's SNAG 1
recommendation (grow TARGET to a file list and keep the entries): the
pipeline rules for this extraction mandate baseline removal without
restructuring the ratchet; the two fields now live outside ratchet
jurisdiction in the new module.
